feat: add new designs for member and spaces profile
The individual profile pages were originally optimised
for workspaces. The changes here make it work better
with our member profile Type.

Closes #1200
