+++ src/components/Form/ImageInput.field.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import * as React from 'react'
 import { ImageInput } from '../ImageInput/ImageInput'
 import { ErrorMessage, FieldContainer } from './elements'
 import { IFieldProps } from './Fields'
@@ -19,27 +19,29 @@ export const ImageInputField = ({
   'data-cy': dataCy,
   customChange,
   ...rest
-}: IExtendedFieldProps) => (
-  <>
-    <FieldContainer
-      style={{ height: '100%', width: '100%', overflow: 'hidden' }}
-      invalid={meta.touched && meta.error}
-      data-cy={dataCy}
-    >
-      <ImageInput
-        {...rest}
-        value={input.value}
-        // as validation happens on blur also want to artificially trigger when values change
-        // (no native blur event)
-        onFilesChange={value => {
-          input.onChange(value)
-          if (customChange) {
-            customChange(value)
-          }
-          input.onBlur()
-        }}
-      />
-    </FieldContainer>
-    {meta.error && meta.touched && <ErrorMessage>{meta.error}</ErrorMessage>}
-  </>
-)
+}: IExtendedFieldProps) => {
+  return (
+    <>
+      <FieldContainer
+        style={{ height: '100%', width: '100%', overflow: 'hidden' }}
+        invalid={meta.touched && meta.error}
+        data-cy={dataCy}
+      >
+        <ImageInput
+          {...rest}
+          value={input.value}
+          // as validation happens on blur also want to artificially trigger when values change
+          // (no native blur event)
+          onFilesChange={value => {
+            input.onChange(value)
+            if (customChange) {
+              customChange(value)
+            }
+            input.onBlur()
+          }}
+        />
+      </FieldContainer>
+      {meta.error && meta.touched && <ErrorMessage>{meta.error}</ErrorMessage>}
+    </>
+  )
+}
