+++ src/pages/Settings/content/formSections/UserInfos.section.tsx
@@ -14,6 +14,8 @@ import { required } from 'src/utils/validators'
 import { IUserPP } from 'src/models/user_pp.models'
 import { ImageInputField } from 'src/components/Form/ImageInput.field'
 import { ErrorMessage } from 'src/components/Form/elements'
+import { IUser } from 'src/models'
+import { IUploadedFileMeta } from 'src/stores/storage'
 
 interface IProps {
   formValues: IUserPP
@@ -26,6 +28,83 @@ interface IState {
   isOpen?: boolean
 }
 
+const CoverImages = ({
+  isMemberProfile,
+  coverImages,
+}: {
+  isMemberProfile: boolean
+  coverImages: IUser['coverImages']
+}) =>
+  isMemberProfile ? (
+    <>
+      <Text mb={2} mt={7} width="100%" medium>
+        Add a profile image *
+      </Text>
+      <Box height="150px" width="150px" m="10px" data-cy="cover-image">
+        <Field
+          hasText={false}
+          name="coverImages[0]"
+          validate={required}
+          validateFields={[]}
+          component={ImageInputField}
+          data-cy={`coverImage`}
+          initialValue={coverImages[0]}
+        />
+      </Box>
+    </>
+  ) : (
+    <>
+      <Text mb={2} mt={7} width="100%" medium>
+        Cover Image *
+      </Text>
+      <FieldArray
+        name="coverImages"
+        initialValue={coverImages as IUploadedFileMeta[]}
+      >
+        {({ fields, meta }) => {
+          return (
+            <>
+              {fields.map((name, index: number) => (
+                <Box
+                  key={name}
+                  height="100px"
+                  width="150px"
+                  m="10px"
+                  data-cy="cover-image"
+                >
+                  <Field
+                    hasText={false}
+                    name={name}
+                    validateFields={[]}
+                    data-cy={`coverImages-${index}`}
+                    component={ImageInputField}
+                  />
+                </Box>
+              ))}
+              {meta.error && <ErrorMessage>{meta.error}</ErrorMessage>}
+            </>
+          )
+        }}
+      </FieldArray>
+
+      <Box
+        bg={theme.colors.softblue}
+        mt={2}
+        p={2}
+        width={1}
+        sx={{ borderRadius: '3px' }}
+      >
+        <Text small>
+          The cover images are shown in your profile and helps us evaluate your
+          account.
+        </Text>
+        <Text small>
+          Make sure the first image shows your space. Best size is 1920x1080.
+        </Text>
+      </Box>
+    </>
+  )
+
 export class UserInfosSection extends React.Component<IProps, IState> {
   constructor(props: IProps) {
     super(props)
@@ -37,8 +116,9 @@ export class UserInfosSection extends React.Component<IProps, IState> {
 
   render() {
     const { formValues } = this.props
-    const { links, coverImages } = formValues
+    const { profileType, links, coverImages } = formValues
     const { isOpen } = this.state
+    const isMemberProfile = profileType === 'member'
     return (
       <FlexSectionContainer>
         <Flex justifyContent="space-between">
@@ -65,7 +145,9 @@ export class UserInfosSection extends React.Component<IProps, IState> {
             />
 
             <Text mb={2} mt={7} medium>
-              Description *
+              {isMemberProfile
+                ? 'Tell us a bit about yourself *'
+                : 'Description *'}
             </Text>
             <Field
               data-cy="info-description"
@@ -75,54 +157,11 @@ export class UserInfosSection extends React.Component<IProps, IState> {
               validate={required}
               validateFields={[]}
             />
-            <Text mb={2} mt={7} width="100%" medium>
-              Cover Image *
-            </Text>
-            <FieldArray name="coverImages" initialValue={coverImages as any}>
-              {({ fields, meta }) => {
-                return (
-                  <>
-                    {fields.map((name, index: number) => (
-                      <Box
-                        key={name}
-                        height="100px"
-                        width="150px"
-                        m="10px"
-                        data-cy="cover-image"
-                      >
-                        <Field
-                          hasText={false}
-                          name={name}
-                          validateFields={[]}
-                          data-cy={`coverImages-${index}`}
-                          component={ImageInputField}
-                        />
-                      </Box>
-                    ))}
-                    {meta.error && <ErrorMessage>{meta.error}</ErrorMessage>}
-                  </>
-                )
-              }}
-            </FieldArray>
-
-            <Box
-              bg={theme.colors.softblue}
-              mt={2}
-              p={2}
-              width={1}
-              sx={{ borderRadius: '3px' }}
-            >
-              <Text small>
-                The cover images are shown in your profile and helps us evaluate
-                your account.
-              </Text>
-              <Text small>
-                Make sure the first image shows your space. Best size is
-                1920x1080.
-              </Text>
-            </Box>
+            <CoverImages
+              isMemberProfile={isMemberProfile}
+              coverImages={coverImages}
+            />
           </Flex>
-
           <>
             <Flex wrap={'nowrap'} alignItems={'center'} width={1}>
               <Text mb={2} mt={7} medium>
