+++ src/pages/User/User.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Route, Switch, withRouter } from 'react-router-dom'
 import { inject, observer } from 'mobx-react'
 import { UserStore } from 'src/stores/User/user.store'
-import { UserPage } from './content/UserPage/UserPage'
+import { UserPage } from './content'
 import { NotFoundPage } from '../NotFound/NotFound'
 
 interface IProps {
